+++ src/test/java/cart/fixture/MemberFixture.java
@@ -9,8 +9,8 @@ public class MemberFixture {
         private static final long ID = 1L;
         private static final String EMAIL = "junho5336@gmail.com";
         private static final String PASSWORD = "1234";
-        private static final Integer MONEY = 1000000;
-        private static final Integer POINT = 5000;
+        private static final long MONEY = 1000000;
+        private static final long POINT = 5000;
 
         public static final Member MEMBER = new Member(ID, EMAIL, PASSWORD, MONEY, POINT);
         public static final MemberEntity ENTITY = new MemberEntity(ID, EMAIL, PASSWORD, MONEY, POINT);
@@ -20,8 +20,8 @@ public class MemberFixture {
         private static final long ID = 2L;
         private static final String EMAIL = "wonny921@gmail.com";
         private static final String PASSWORD = "1234";
-        private static final Integer MONEY = 1000000;
-        private static final Integer POINT = 5000;
+        private static final long MONEY = 1000000;
+        private static final long POINT = 5000;
 
         public static final Member MEMBER = new Member(ID, EMAIL, PASSWORD, MONEY, POINT);
         public static final MemberEntity ENTITY = new MemberEntity(ID, EMAIL, PASSWORD, MONEY, POINT);
@@ -31,8 +31,8 @@ public class MemberFixture {
         private static final long ID = 3L;
         private static final String EMAIL = "goatsfish@gmail.com";
         private static final String PASSWORD = "1234";
-        private static final Integer MONEY = 1000000;
-        private static final Integer POINT = 5000;
+        private static final long MONEY = 1000000;
+        private static final long POINT = 5000;
 
         public static final Member MEMBER = new Member(ID, EMAIL, PASSWORD, MONEY, POINT);
         public static final MemberEntity ENTITY = new MemberEntity(ID, EMAIL, PASSWORD, MONEY, POINT);
@@ -42,8 +42,8 @@ public class MemberFixture {
         private static final long ID = 4L;
         private static final String EMAIL = "dani@gmail.com";
         private static final String PASSWORD = "12345";
-        private static final Integer MONEY = 1000000;
-        private static final Integer POINT = 5000;
+        private static final long MONEY = 1000000;
+        private static final long POINT = 5000;
 
         public static final Member MEMBER = new Member(ID, EMAIL, PASSWORD, MONEY, POINT);
         public static final MemberEntity ENTITY = new MemberEntity(ID, EMAIL, PASSWORD, MONEY, POINT);
@@ -53,8 +53,8 @@ public class MemberFixture {
         private static final long ID = 5L;
         private static final String EMAIL = "oing@gmail.com";
         private static final String PASSWORD = "12345";
-        private static final Integer MONEY = 0;
-        private static final Integer POINT = 10000;
+        private static final long MONEY = 0;
+        private static final long POINT = 10000;
 
         public static final Member MEMBER = new Member(ID, EMAIL, PASSWORD, MONEY, POINT);
         public static final MemberEntity ENTITY = new MemberEntity(ID, EMAIL, PASSWORD, MONEY, POINT);
@@ -65,8 +65,8 @@ public class MemberFixture {
         private static final long ID = 6L;
         private static final String EMAIL = "oingoing@gmail.com";
         private static final String PASSWORD = "12345";
-        private static final Integer MONEY = 100000;
-        private static final Integer POINT = 0;
+        private static final long MONEY = 100000;
+        private static final long POINT = 0;
 
         public static final Member MEMBER = new Member(ID, EMAIL, PASSWORD, MONEY, POINT);
         public static final MemberEntity ENTITY = new MemberEntity(ID, EMAIL, PASSWORD, MONEY, POINT);
